+++ src/db/redis_db.cc
@@ -216,7 +216,7 @@ int32_t redis_db::print_all_keys(void)
 //: \return:  TODO
 //: \param:   TODO
 //: ----------------------------------------------------------------------------
-int32_t redis_db::set_opt(uint32_t a_opt, const void *a_buf, uint32_t a_len)
+int32_t redis_db::set_opt(uint32_t a_opt, const void *a_buf, uint64_t a_len)
 {
         switch(a_opt)
         {
